% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infx.R
\name{as_list}
\alias{as_list}
\title{Alias for \code{as.list()}}
\usage{
as_list(x, ...)
}
\description{
For consistency with other functions provided by this package, where for
example both \code{is_json_vec()} and \code{is.json_vec()} are available, \code{as_list()}
is provided as an alias for \code{as.list()}. Consequently, functions such as
\code{as.list.json_vec()} are also available as \code{as_list.json_vec()}.
}
