% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_dataset_id}
\alias{create_dataset_id}
\title{Create data set id objects}
\usage{
create_dataset_id(token, data_set)
}
\arguments{
\item{token}{Login token as created by \link{login_openbis}.}

\item{data_set}{Either a single/list of json_class object(s), each
containing a datasetCode field or a character vector.}
}
\value{
A single or a list of DatasetIdentifier object(s).
}
\description{
From a list of objects, each containing a datasetCode field or
from a character vector of datasetCodes, create a list of DatasetIdentifier
objects.
}
