% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_exp_ids}
\alias{create_exp_ids}
\title{Create experiment id objects}
\usage{
create_exp_ids(exp_code = NULL, proj_code = NULL, space_code = NULL,
  perm_id = NULL, token = NULL, ...)
}
\arguments{
\item{exp_code, proj_code, space_code, perm_id}{If all are not NULL, character
vectors of length 1 that make up the returned ExperimentIdentifier object
or if one or more is NULL, regular expressions to match fields against the
full list of available ExperimentIdentifier objects.}

\item{token}{Login token as created by \link{login_openbis}.}

\item{...}{Passed to \link{grepl}.}
}
\value{
A single or a list of ExperimentIdentifier object(s).
}
\description{
Either an ExperimentIdentifier object is created by specifying
all needed information or, if some information is missing, openBis is
queried for all available ExperimentIdentifiers and the provided regular
expressions are used to select a subset of all ids.
}
