% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_plate_id}
\alias{create_plate_id}
\title{Create a plate id object}
\usage{
create_plate_id(plate_id, space_code = NULL, token = NULL)
}
\arguments{
\item{plate_id}{Plate barcode.}

\item{space_code}{The space code of the plate; it NULL, it is determined
automatically.}

\item{token}{Login token as created by [login_openbis].}
}
\value{
A single PlateIdentifier object as a list of type json_class.
}
\description{
PlateIdentifier objects are used throughout the openBis API
to identify plates. In addition to the plate barcode, the space the plate
resides in is needed. If unknown (NULL), this property can be determined by
listing all Plates (using [list_plates]) and matching the provided barcode.
If several plates in different spaces have the same barcode, an error is
thrown.
}
