% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_code}
\alias{dataset_code}
\alias{dataset_code.DataSet}
\alias{dataset_code.DatasetIdentifier}
\alias{dataset_code.DatasetReference}
\alias{dataset_code.FeatureVectorDatasetReference}
\alias{dataset_code.FeatureVectorDatasetWellReference}
\alias{dataset_code.ImageDatasetReference}
\alias{dataset_code.MicroscopyImageReference}
\alias{dataset_code.PlateImageReference}
\alias{dataset_code.DataSetFileDTO}
\title{Extract dataset code}
\usage{
dataset_code(x, ...)

\method{dataset_code}{DataSet}(x, ...)

\method{dataset_code}{DatasetIdentifier}(x, ...)

\method{dataset_code}{DatasetReference}(x, ...)

\method{dataset_code}{FeatureVectorDatasetReference}(x, ...)

\method{dataset_code}{FeatureVectorDatasetWellReference}(x, ...)

\method{dataset_code}{ImageDatasetReference}(x, ...)

\method{dataset_code}{MicroscopyImageReference}(x, ...)

\method{dataset_code}{PlateImageReference}(x, ...)

\method{dataset_code}{DataSetFileDTO}(x, ...)
}
\arguments{
\item{x}{Dataset object(s).}

\item{...}{Generic compatibility.}
}
\description{
Given a (set of) dataset object(s), for each one extract the dataset code
and return a character vector of codes.
}
\keyword{internal}
