% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{do_download}
\alias{do_download}
\title{Download files}
\usage{
do_download(token, data_id, files, rep = 1, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{data_id}{Data set code.}

\item{files}{Files objects as produced by [list_files] describing the files
to be downloaded.}

\item{rep}{The number of times failed downloads are repeated.}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
A list of raw vectors holding the downloaded data.
}
\description{
Given a data set code and a set of file paths (as returned
from [list_files]), a download link is generated for each file by
[get_download]. The downloads are preformed asynchronously.
}
