% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbis.R
\name{do_openbis}
\alias{do_openbis}
\title{Helper querying openBis API}
\usage{
do_openbis(method, params,
  api = "openbis/openbis/rmi-general-information-v1.json",
  url = "https://infectx.biozentrum.unibas.ch")
}
\arguments{
\item{method}{The method name}

\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The "@type" entries will be
generated from "json_class" attributes.}

\item{api}{The location of the JS libraries handling the request. Is
appended to the supplied url.}

\item{url}{The base url, the request is sent to.}
}
\value{
A list/data.frame holding the response from openBis.
}
\description{
Issues a POST request to the JSON-RPC based openBis API v1.
Documentation is available at \code{https://wiki-bsse.ethz.ch/display/
openBISDoc/openBIS+JSON+API}.
}
