% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_default}
\alias{error_default}
\title{Helper for S3 default error}
\usage{
error_default(x, ...)
}
\arguments{
\item{x}{The object on which the dispatch is done.}
}
\description{
Whenever the default function of an S3 generic is a \code{not implemented} error,
tis function can be used.
}
\keyword{internal}
