% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{exp_id_str}
\alias{exp_id_str}
\alias{exp_id_str.ExperimentIdentifier}
\alias{exp_id_str.Experiment}
\title{Extract experiment string}
\usage{
exp_id_str(x, ...)

\method{exp_id_str}{ExperimentIdentifier}(x, ...)

\method{exp_id_str}{Experiment}(x, ...)
}
\arguments{
\item{x}{Experiment object(s).}

\item{...}{Generic compatibility.}
}
\description{
Experiments can be uniquely identified by a string made up of space code,
project code and experiment code. This function extracts the relevant fields
from experiment or experiment id objects and returns a vector of experiment
strings.
}
\keyword{internal}
