% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Extract DataSets from a list}
\usage{
extract_dataset(ds, type = NULL, most_recent = FALSE)
}
\arguments{
\item{ds}{The list of DataSets to operate on.}

\item{type}{A string holding the type to be matched.}

\item{most_recent}{Logical switch whether to select the most recent DataSet.}
}
\value{
A (subsetted) list of DataSets.
}
\description{
Finds and extracts the DataSets from a list of DataSets that
match the argument type and/or are the most recent available.
}
