% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feat_ds_well_ref}
\alias{feat_ds_well_ref}
\title{Create well feature dataset reference objects}
\usage{
feat_ds_well_ref(x, wells)
}
\arguments{
\item{x}{A set of \code{FeatureVectorDatasetReference} objects.}

\item{wells}{A set of \code{WellPosition} objects.}
}
\description{
While \code{FeatureVectorDatasetReference} objects represent feature datasets on
a plate level, \code{FeatureVectorDatasetWellReference} reference feature
datasets on well-level. In order to create such well-level representations,
a set of \code{FeatureVectorDatasetReference} and a set of \code{WellPosition} are
combined to \code{FeatureVectorDatasetWellReference} objects where each instance
contains a single object of the inputted sets. If either argument is of
length greater than 1, the other argument has to be of the same length or of
length 1, in which case it will be \code{\link[=base:rep]{base:rep()}}eated to the required length.
}
\keyword{internal}
