% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from openBis.}
\usage{
fetch_data(auto_rm = TRUE, ...)
}
\arguments{
\item{auto_rm}{A logical switch; if TRUE, all downloaded files will be
removed upon garbage collection of the returned character vector.}

\item{...}{Arguments passed to [fetch_openbis()]}
}
\value{
A vector of downloaded filenames.
}
\description{
Calling [fetch_openbis()], this function downloads data and
returns the downloaded filenames. It optionally makes sure the data is
removed again upon garbage collection of the returned filename reference.
}
