% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{fetch_features}
\alias{fetch_features}
\alias{fetch_features.FeatureVectorDatasetReference}
\alias{fetch_features.FeatureVectorDatasetWellReference}
\title{Fetch feature data}
\usage{
fetch_features(token, x, feature_codes = NA, ...)

\method{fetch_features}{FeatureVectorDatasetReference}(token, x,
  feature_codes = NA, ...)

\method{fetch_features}{FeatureVectorDatasetWellReference}(token, x,
  feature_codes = NA, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to specify the set of feature vector data sets of interest.}

\item{feature_codes}{A character vector of feature codes or NA (all
available feature codes).}

\item{...}{Generic compatibility}
}
\description{
For a given set of feature vector data sets, \code{fetch_features()} fetches
feature data for the specified feature codes, or for all available features
in case the argument \code{feature_codes} is not specified (or \code{NA}). If
for different data sets different sets of features are available,
the union of the features of all data sets is searched for. The returned
object is of type \code{FeatureVectorDataset}, which for each entry contains a
\code{FeatureVectorDatasetReference} and a set of \code{FeatureVector}(s), one for
each well.
}
\details{
If only a limited set of wells is of interest, the search can be limited by
dispatching \code{fetch_features()} on a (set of)
\code{FeatureVectorDatasetWellReference} objects.
}
\section{TODO}{
 Even though there exists a constructor for
\code{FeatureVectorDatasetWellReference} objects, which takes two arguments, one
for the corresponding \code{FeatureVectorDatasetReference} object and one for
a \code{WellPosition} objects, this does not work. Furthermore, class information
cannot be supplied as this will cause an error as well (hence the use of
\code{rm_json_class()}). Why the function \code{loadFeaturesForDatasetWellReferences}
behaves this way is currently unclear.
}

