% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R, R/file.R
\name{fetch_files}
\alias{fetch_files}
\alias{fetch_files}
\alias{fetch_files.FileInfoDssDTO}
\alias{fetch_files_serial}
\title{Split downloads}
\usage{
fetch_files(token, x, ...)

fetch_files(token, x, ...)

\method{fetch_files}{FileInfoDssDTO}(token, x, data_sets, ...)

fetch_files_serial(urls, retry = 2L, file_sizes = NULL, done = identity,
  ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to specify which files to download.}

\item{...}{Generic compatibility.}

\item{data_sets}{Either a single dataset object (anything that has a
\code{dataset_code()} method) or a set of objects of the same length as \code{x}.}

\item{urls}{Either a caracter vector or a list of calls that each yields an
url when \code{eval}d.}

\item{retry}{The number of tries for each url.}

\item{file_sizes}{A vector of expected file sizes or NULL.}

\item{done}{A function with a single argument which is applied to each
downloaded file.}

\item{async}{The bucket size for simultaneous downloads.}
}
\value{
A list of raw vectors holding the downloaded data.
}
\description{
A set of files is split into buckets and downloaded
asynchronously and the progress is reported using a progress bar. The
splitting is necessary as \link{do_download} downloads all files simultaneously.

The function \code{fetch_files()} downloads files specified by a set of
\code{FileInfoDssDTO} objects alongside one or several dataset objects.
}
\section{TODO}{
 entry point to all downloads, choose single or async (both
not exported) and use this in plate/meta fetching.
}

