% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{fetch_images}
\alias{fetch_images}
\alias{fetch_images.DatasetIdentifier}
\alias{fetch_images.DatasetReference}
\alias{fetch_images.ImageDatasetReference}
\alias{fetch_images.MicroscopyImageReference}
\alias{fetch_images.PlateImageReference}
\title{Fetch images}
\usage{
fetch_images(token, x, ...)

\method{fetch_images}{DatasetIdentifier}(token, x, channels,
  well_positions = NULL, image_size = NULL, ...)

\method{fetch_images}{DatasetReference}(token, x, channels,
  well_positions = NULL, image_size = NULL, ...)

\method{fetch_images}{ImageDatasetReference}(token, x, channels,
  well_positions = NULL, image_size = NULL, ...)

\method{fetch_images}{MicroscopyImageReference}(token, x,
  well_positions = NULL, image_size = NULL, ...)

\method{fetch_images}{PlateImageReference}(token, x, image_size = NULL,
  force_png = FALSE, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned images}

\item{...}{Generic compatibility}

\item{channels}{A character vector of imaging channels}

\item{well_positions}{A (list of) \code{WellPosition} objects. If the object
passed as argument x already contains well position information this can
be NULL.}

\item{image_size}{Either a single \code{ImageSize} object or NULL, in which case
images are returned in full size.}

\item{force_png}{Logical switch for making sure the returned image is a
png. If NULL or FALSE, the image is returned in the format it is stored.}
}
\description{
Download Base64 encoded images as strings.
}
