% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_meta}
\alias{fetch_meta}
\title{Fetch InfectX meta data}
\usage{
fetch_meta(type = c("db", "public"), ...)
}
\arguments{
\item{type}{A switch for the type of meta data to be downloaded.}

\item{...}{Passed to [fetch_data()].}
}
\value{
A vector of downloaded filenames
}
\description{
A wrapper around [fetch_data()], this function essentially
provides default values for downloading meta data from openBis. Two formats
of meta data are currently available: dumps of SQLite databased holding the
entire set of experimental meta data, as well as a CSV sheet, containing
only the published subset.
}
