% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_openbis}
\alias{fetch_openbis}
\title{Download datasets from an openBis instance}
\usage{
fetch_openbis(username = "rdgr2014", password = "IXPubReview",
  data_type = NULL, data_id = NULL, plate_regex = NULL,
  file_regex = ".*", out_dir = getwd(), result_class = "stable",
  newest = TRUE, verbosity = 6)
}
\arguments{
\item{username, password}{OpneBis login credentials}

\item{data_type}{Specify the type of dataset to be downloaded}

\item{data_id}{If applicable, for example for a single aggregate file,
directly select the single dataset to download (optional)}

\item{plate_regex}{If datasets corresponding to one or several plates are
targeted, they can be selected via a regular expression (optional)}

\item{file_regex}{A regular expression applied to filenames can be used to
filter the selection (optional)}

\item{out_dir}{Directory used to save the downloaded data to (default: the
current working dir)}

\item{result_class}{A string, indicating the result class to restrict to.
NULL to skip.}

\item{newest}{Logical, indicating whether to only return the newest
version as a matched dataset per plate (default is TRUE).}

\item{verbosity}{The verbosity level of BeeDataSetDownloader (an integer in
[0-25])}
}
\value{
A character vector holding all std output of BeeDataSetDownloader at
the specified verbosity level.
}
\description{
This is a wrapper for the BeeDataSetDownloader.jar application
supplied by the package, responsible for downloading data from the InfectX
openBis instance.
}
