% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbis.R
\name{fetch_plate}
\alias{fetch_plate}
\title{Fetch single cell data}
\usage{
fetch_plate(plate_name, auto_rm = TRUE, ...)
}
\arguments{
\item{plate_name}{The plate barcode of interest.}

\item{auto_rm}{A logical switch; if TRUE, all downloaded files will be
removed upon garbage collection of the returned character vector.}

\item{...}{Arguments passed to [fetch_openbis()]}
}
\value{
A vector of downloaded filenames.
}
\description{
Using [fetch_openbis()], this function downloads single cell
datasets corresponding to a plate barcode and filtered by an optional file
name regular expression. The downloaded data is either saved to a user
specified directory or to a temporary location which can be auto removed.
}
