% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{has_json_class}
\alias{has_json_class}
\title{Test if object has a JSON class}
\usage{
has_json_class(x, class = NULL)
}
\arguments{
\item{x}{Object to test.}

\item{class}{(Optional) class name to test.}
}
\value{
A single logical.
}
\description{
Either tests whether an object has any JSON class attached or
a specific one.
}
