% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-class.R
\name{print_json_class}
\alias{print_json_class}
\alias{style}
\title{Helper functions for printing JSON objects}
\usage{
print_json_class(x, cur_depth, max_depth, layout = style())

style(fancy = TRUE)
}
\arguments{
\item{x}{The JSON object to print.}

\item{cur_depth}{The current recursion depth.}

\item{max_depth}{The maximum recursion depth.}

\item{layout}{Characters for printing the tree structure and styles to be
applied to the different entities.}

\item{fancy}{Logical switch to enable font styles, colors and UTF box
characters.}
}
\description{
Inspired by the ast printing function of Hadley's \code{lobstr} package heavily
borrowing code from \href{https://git.io/vFMA5}{there}, this enables the
recursive printing of JSON objects. Printing style can either be with fancy
(colors, UTF box characters. etc.) or simple.
}
\section{todo Make sure that structures such as lst print correctly}{
NA
}

\keyword{internal}
