% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-vec.R
\name{json_vec}
\alias{json_vec}
\alias{new_json_vec}
\alias{as.json_vec}
\alias{as.json_vec.json_vec}
\alias{as.json_vec.json_class}
\alias{as.json_vec.list}
\alias{as.json_vec.default}
\alias{as.list.json_vec}
\alias{has_common_subclass}
\alias{get_common_subclass}
\alias{get_common_subclass.json_class}
\alias{get_common_subclass.list}
\alias{get_common_subclass.json_vec}
\alias{get_common_subclass.default}
\title{Construct a json_vec object}
\usage{
json_vec(...)

new_json_vec(x)

as.json_vec(x, ...)

\method{as.json_vec}{json_vec}(x, ...)

\method{as.json_vec}{json_class}(x, ...)

\method{as.json_vec}{list}(x, ...)

\method{as.json_vec}{default}(x, ...)

\method{as.list}{json_vec}(x, ...)

has_common_subclass(x)

get_common_subclass(x, ...)

\method{get_common_subclass}{json_class}(x, ...)

\method{get_common_subclass}{list}(x, ...)

\method{get_common_subclass}{json_vec}(x, ...)

\method{get_common_subclass}{default}(x, ...)
}
\arguments{
\item{...}{Individual \code{json_class} objects}

\item{x}{A single/list of \code{json_class} object(s), or other object to coerce}
}
\description{
\code{new_json_vec} is a low-level constructor that takes a list of
\code{json_class} objects of the same sub-class. \code{json_vec} constructs a json_vec
from individual json_class objects and \code{as.json_vec} is a S3 generic that
converts existing objects.
}
\examples{
a <- structure("a", class = c("foo", "json_class"))
b <- structure("b", class = c("foo", "json_class"))

new_json_vec(list(a, b))
json_vec(a, b)

as.json_vec(list(a, b))

}
