% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_dataset_types}
\alias{list_dataset_types}
\title{List dataset types}
\usage{
list_dataset_types(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}
}
\value{
A a list of DataSetType objects.
}
\description{
Given a login token, all available dataset types are listed.
}
