% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{list_datastores}
\alias{list_datastores}
\alias{list_datastore_urls}
\title{List data store servers}
\usage{
list_datastores(token)

list_datastore_urls(token, data_set = NULL)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{data_set}{Set of datasets (specified as dataset codes, \code{DataSet} or
\code{DatasetIdentifier} objects) or NULL.}
}
\description{
\code{list_datastores()} lists all data store servers registered this openBIS
server instance and \code{list_datastore_urls()} returns the base urls of the
data store servers that host the supplied datasets (or that of the default
data store server in case no data set is supplied).
}
