% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_exp_datasets}
\alias{list_exp_datasets}
\title{Get data sets for a set of experiments}
\usage{
list_exp_datasets(token, experiment)
}
\arguments{
\item{token}{Login token as created by \link{login_openbis}.}

\item{experiment}{A data.frame representing a set of experiments.}
}
\value{
List, containing (among others), entries "code" and
"dataSetTypeCode".
}
\description{
Given a set of experiments as a data.frame, the corresponding
data sets are queried. Can be slow if several experiments are included in
the query.
}
