% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{list_experiment_types}
\alias{list_experiment_types}
\title{List experiment types}
\usage{
list_experiment_types(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}
}
\value{
A a list of ExperimentType objects.
}
\description{
Given a login token, all available experiment types are listed.
}
