% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{list_features}
\alias{list_features}
\alias{list_features.FeatureVectorDatasetReference}
\alias{list_features.FeatureVectorDatasetWellReference}
\alias{list_feature_codes}
\alias{list_feature_codes.FeatureVectorDatasetReference}
\alias{list_feature_codes.FeatureVectorDatasetWellReference}
\title{List features}
\usage{
list_features(token, x, feature_codes = NA, ...)

\method{list_features}{FeatureVectorDatasetReference}(token, x,
  feature_codes = NA, ...)

\method{list_features}{FeatureVectorDatasetWellReference}(token, x,
  feature_codes = NA, ...)

list_feature_codes(token, x, ...)

\method{list_feature_codes}{FeatureVectorDatasetReference}(token, x, ...)

\method{list_feature_codes}{FeatureVectorDatasetWellReference}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to specify the set of feature vector data sets of interest.}

\item{...}{Generic compatibility}
}
\description{
For a given set of feature vector data sets, \code{list_features()} provides the
list of all available features. This contains the code, label and
description of the feature, as (set of) \code{FeatureInformation} object(s). If
for different data sets different sets of features are available,
\code{list_features()} provides the union of the features of all data sets.
}
\details{
Similarly, \code{list_feature_codes()} provides the list of all available
features as character vector or feature codes. If for different data sets
different sets of features are available, \code{list_feature_codes()} provides
the union of the features of all data sets.
}
