% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{list_features}
\alias{list_features}
\alias{list_features.FeatureVectorDatasetReference}
\alias{list_features.FeatureVectorDatasetWellReference}
\alias{list_feature_codes}
\alias{list_feature_codes.FeatureVectorDatasetReference}
\alias{list_feature_codes.FeatureVectorDatasetWellReference}
\alias{fetch_features}
\alias{fetch_features.FeatureVectorDatasetReference}
\alias{fetch_features.FeatureVectorDatasetWellReference}
\title{List and download feature data}
\usage{
list_features(token, x, ...)

\method{list_features}{FeatureVectorDatasetReference}(token, x, wells = NULL,
  ...)

\method{list_features}{FeatureVectorDatasetWellReference}(token, x, ...)

list_feature_codes(token, x, ...)

\method{list_feature_codes}{FeatureVectorDatasetReference}(token, x,
  wells = NULL, ...)

\method{list_feature_codes}{FeatureVectorDatasetWellReference}(token, x, ...)

fetch_features(token, x, feature_codes = NA, ...)

\method{fetch_features}{FeatureVectorDatasetReference}(token, x,
  feature_codes = NA, wells = NULL, ...)

\method{fetch_features}{FeatureVectorDatasetWellReference}(token, x,
  feature_codes = NA, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to specify the set of feature vector datasets of interest.}

\item{...}{Generic compatibility}

\item{feature_codes}{A character vector of feature codes or NA (all
available feature codes).}
}
\description{
In openBIS, features are datasets that are treated differently from
generic datasets. Briefly put, tabular datasets where columns correspond
to features and rows to wells can be marked as feature datasets which makes
it possible to query openBIS for individual feature values for selected
wells instead of having to download the entire table for a plate. The
relevant object types for handling feature data are
\code{FeatureVectorDatasetReference} and \code{FeatureVectorDatasetWellReference},
where the former represents feature data at plate level and the latter at
well level. The function \code{list_features()} can be used to enumerate
available features and \code{fetch_features()} will download feature data.
}
\details{
For a given set of feature vector datasets, \code{list_features()} provides the
list of all available features. This contains the code, label and
description of the feature, as (set of) \code{FeatureInformation} object(s). If
for different datasets different sets of features are available,
\code{list_features()} provides the union of the features of all datasets.

Similarly, \code{list_feature_codes()} provides the list of all available
features as character vector or feature codes. If for different datasets
different sets of features are available, \code{list_feature_codes()} provides
the union of the features of all datasets.

For a given set of feature vector datasets, \code{fetch_features()} fetches
feature data for the specified feature codes, or for all available features
in case the argument \code{feature_codes} is not specified (or \code{NA}). If
for different datasets different sets of features are available,
the union of the features of all datasets is searched for. The returned
object is of type \code{FeatureVectorDataset}, which for each entry contains a
\code{FeatureVectorDatasetReference} and a set of \code{FeatureVector}(s), one for
each well.

If only a limited set of wells is of interest, the search can be limited by
dispatching \code{fetch_features()} on a (set of)
\code{FeatureVectorDatasetWellReference} objects.
}
\section{openBIS}{

\itemize{
\item \Sexpr{infx::docs_link("dsrs", "listAvailableFeatures")}
}


\itemize{
\item \Sexpr{infx::docs_link("dsrs", "listAvailableFeatureCodes")}
}


\itemize{
\item \Sexpr{infx::docs_link("dsrs", "loadFeatures")}
}


\itemize{
\item \Sexpr{infx::docs_link("dsrs", "loadFeaturesForDatasetWellReferences")}
}
}

\section{TODO}{
 Even though there exists a constructor for
\code{FeatureVectorDatasetWellReference} objects, which takes two arguments, one
for the corresponding \code{FeatureVectorDatasetReference} object and one for
a \code{WellPosition} objects, this does not work. Furthermore, class information
cannot be supplied as this will cause an error as well (hence the use of
\code{rm_json_class()}). Why the function \code{loadFeaturesForDatasetWellReferences}
behaves this way is currently unclear.
}

