% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{fetch_images}
\alias{fetch_images}
\alias{fetch_images.DatasetIdentifier}
\alias{fetch_images.DatasetReference}
\alias{fetch_images.ImageDatasetReference}
\alias{fetch_images.MicroscopyImageReference}
\alias{fetch_images.PlateImageReference}
\alias{list_image_metadata}
\alias{list_image_metadata.ExperimentIdentifier}
\alias{list_image_metadata.Experiment}
\alias{list_image_metadata.DatasetIdentifier}
\alias{list_image_metadata.DatasetReference}
\alias{list_image_metadata.ImageDatasetReference}
\alias{list_image_metadata.MicroscopyImageReference}
\alias{list_image_metadata.PlateImageReference}
\title{List image meta data and download images}
\usage{
fetch_images(token, x, ...)

\method{fetch_images}{DatasetIdentifier}(token, x, channels,
  well_positions = NULL, image_size = NULL, thumbnails = FALSE, ...)

\method{fetch_images}{DatasetReference}(token, x, channels,
  well_positions = NULL, image_size = NULL, thumbnails = FALSE, ...)

\method{fetch_images}{ImageDatasetReference}(token, x, channels,
  well_positions = NULL, image_size = NULL, thumbnails = FALSE, ...)

\method{fetch_images}{MicroscopyImageReference}(token, x,
  well_positions = NULL, image_size = NULL, thumbnails = FALSE, ...)

\method{fetch_images}{PlateImageReference}(token, x, image_size = NULL,
  force_png = FALSE, format = NULL, thumbnails = FALSE, ...)

list_image_metadata(token, x, ...)

\method{list_image_metadata}{ExperimentIdentifier}(token, x, ...)

\method{list_image_metadata}{Experiment}(token, x, ...)

\method{list_image_metadata}{DatasetIdentifier}(token, x,
  type = c("metadata", "format"), ...)

\method{list_image_metadata}{DatasetReference}(token, x,
  type = c("metadata", "format"), ...)

\method{list_image_metadata}{ImageDatasetReference}(token, x,
  type = c("metadata", "format"), ...)

\method{list_image_metadata}{MicroscopyImageReference}(token, x,
  type = c("metadata", "format"), ...)

\method{list_image_metadata}{PlateImageReference}(token, x,
  type = c("metadata", "format"), ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned images}

\item{...}{Generic compatibility. Extra arguments will be passed to
\code{\link[=make_requests]{make_requests()}}.}

\item{channels}{A character vector of imaging channels}

\item{well_positions}{A (list of) \code{WellPosition} objects. If the object
passed as argument x already contains well position information this can
be NULL.}

\item{image_size}{Either a single \code{ImageSize} object or NULL, in which case
images are returned in full size.}

\item{thumbnails}{Logical switch; if TRUE, thumbnail images are retrieved
in which case the arguments \code{well_positions} and \code{image_size} are expected
to be at their default values.}

\item{force_png}{Logical switch for making sure the returned image is a
png. If NULL or FALSE, the image is returned in the format it is stored.}

\item{format}{If not NULL, a single \code{ImageRepresentationFormat} object.
Cannot be combined with non-default \code{image_size}, \code{force_png} and \code{format}
arguments.}

\item{type}{Switch to specify the type of meta data objects to be returned.}
}
\description{
Experiment level image meta data can be listed by passing experiment
representing objects (\code{Experiment} or \code{ExperimentIdentifier}) to
\code{list_image_metadata()} and data set level image meta data can be retrieved
by passing data set identifying objects which can be associated with image
data sets (data set id and data set reference objects). Images themselves
can be retrieved using \code{fetch_images()}. As with meta data listing, this
function can be dispatched on objects referencing or identifying data sets
associated with image data.
}
\details{
Data set level image meta data can be listed by passing objects, which
implement the \code{IDatasetIdentifier} interface and are connected to image
data sets (this rules out feature data set references and leaves
\code{DatasetIdentifier}, \code{DatasetReference}, \code{ImageDatasetReference},
\code{MicroscopyImageReference} and \code{PlateImageReference} objects). Two
different types of meta data objects are returned, depending on the \code{type}
argument: if it is set to \code{metadata} (default), objects of type
\code{ImageDatasetMetadata} and if it is set to \code{format}, objects of type
\code{DatasetImageRepresentationFormats} are returned. For experiment-level
image meta data, \code{ExperimentImageMetadata} objects are returned.

Dispatch of \code{fetch_images()} is available for the same object types as data
set-level image meta data listing: \code{DatasetIdentifier}, \code{DatasetReference},
\code{ImageDatasetReference}, \code{MicroscopyImageReference} and
\code{PlateImageReference}. The highest level of control over which images are
retrieved is achieved with \code{PlateImageReference} objects, which specify an
image data set, a well, a tile and a channel. The returned image format can
be modified by either passing an \code{ImageRepresentationFormat} object as the
\code{format} argument, by passing a single/list of format selection criterion
objects, which will be used to filter the available image representation
format objects or by specifying one or both of the \code{image_size} (expects an
\code{ImageSize} object) and \code{force_png} (logical switch) arguments.

\code{MicroscopyImageReference} objects contain channel information (as well as
tile information, which is not taken into account though). Therefore a
(list of) \code{WellPosition} object(s) has/have to be specified, for which then
all tiles are fetched for the given imaging channel. If the passed list of
\code{MicroscopyImageReference} objects contain instances that only differ in
tile number, redundancies are filtered out. An API call is necessary for
each non-redundant object.

Finally, \code{DatasetIdentifier}, \code{DatasetReference} and \code{ImageDatasetReference}
objects are all handled identically. For each of the specified data sets,
an imaging channel has to be provided and whenever the data set is
associated with an entire plate, a (list of) \code{WellPosition} object(s) as
well. If the data set is associated with a single well, the
\code{well_positions} can be left at its default value (NULL). If several data
sets are passed, an API call is necessary per data set. Possible
redundancies are not filtered.

Images are retrieved as Base64 encoded strings, which are converted to
binary using \code{\link[base64enc:base64decode]{base64enc::base64decode()}} and subsequently read by
\code{\link[magick:image_read]{magick::image_read()}}. Attached to the images is the information, based
on which they were retrieved, including data set object, well positions
(where applicable) and channel (where applicable). This results in a list
with length corresponding to the number of API calls that were necessary.
}
\section{Implementation notes}{

\itemize{
\item For dispatch on \code{PlateImageReference} objects, currently the only options
controlling the returned images are an argument for image size and a flag
for forcing the returned format to png. OpenBIS also supports
pre-defined image transformations to be applied to the images before they
are sent to the requesting party. These transformations can be requested
by a code (options are listed in \code{ImageRepresentationFormat} objects or
in \code{ImageChannel} objects attached to \code{ImageDatasetMetadata} objects).
However, as no such transformations appear to be defined, this is
currently not implemented.
\item When filtering \code{ImageRepresentationFormat} objects associated with a data
set, only \code{SizeCriterion} objects can be used. The remaining criteria
(\code{ColorDepthCriterion}, \code{FileTypeCriterion} and \code{OriginalCriterion}) are
currently disabled as they extend the abstract class
\code{AbstractFormatSelectionCriterion}, which causes an issue with JSON
deserialization.
}
}

\section{openBIS}{

\itemize{
\item \Sexpr[results=rd]{infx::docs_link("dsrs", "loadImagesBase64")}
\item \Sexpr[results=rd]{infx::docs_link("dsrs", "loadThumbnailImagesBase64")}
}


\itemize{
\item \Sexpr[results=rd]{infx::docs_link("dsrs",
"loadPhysicalThumbnailsBase64")}
}


\itemize{
\item \Sexpr[results=rd]{infx::docs_link("sas", "getExperimentImageMetadata")}
\item \Sexpr[results=rd]{infx::docs_link("dsrs", "listImageMetadata")}
\item \Sexpr[results=rd]{infx::docs_link("dsrs",
"listAvailableImageRepresentationFormats")}
}
}

\examples{
\donttest{
  tok <- login_openbis()

  # search for a sample object corresponding to plate KB2-03-1I
  samp <- search_openbis(tok,
                         search_criteria(
                           attribute_clause("code",
                                            "/INFECTX_PUBLISHED/KB2-03-1I")
                         ),
                         target_object = "sample")
  # for the plate sample object, list raw image data set references
  ds_ref <- list_references(tok, samp)

  # the returned image dataset reference can be used to list image meta data
  img_meta <- list_image_metadata(tok, ds_ref)
  channels <- img_meta[["channelCodes"]]

  imgs <- fetch_images(tok, ds_ref,
                       channels = channels[[1L]],
                       well_positions = well_pos(1, 1),
                       image_size = json_class(width = 300, height = 300,
                                               class = "ImageSize"))
  # this yields 9 images, one per tile
  length(imgs[[1L]]) == img_meta[["numberOfTiles"]]
  # and each image is scaled to fit within 300 x 300 pixels
  magick::image_info(imgs[[1L]][[1L]])

  # if not the entire well is of interest, but only certain tiles
  img_ref <- list_references(tok, ds_ref,
                             wells = well_pos(1, 1),
                             channels = channels[[1L]])
  # this yields 9 objects, one reference per tile
  length(img_ref)
  # select a tile, for example the center one
  img <- fetch_images(tok, img_ref[[5L]],
                      image_size = json_class(width = 300, height = 300,
                                              class = "ImageSize"))
  identical(as.raster(img[[1L]]), as.raster(imgs[[1L]][[5L]]))

  logout_openbis(tok)
}

}
\seealso{
Other resource listing/downloading functions: \code{\link{list_download_urls}},
  \code{\link{list_features}}, \code{\link{list_files}}
}
\concept{resource listing/downloading functions}
