% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{list_image_metadata}
\alias{list_image_metadata}
\alias{list_image_metadata.ExperimentIdentifier}
\alias{list_image_metadata.Experiment}
\alias{list_image_metadata.ImageDatasetReference}
\title{List image metadata}
\usage{
list_image_metadata(token, x, ...)

\method{list_image_metadata}{ExperimentIdentifier}(token, x, ...)

\method{list_image_metadata}{Experiment}(token, x, ...)

\method{list_image_metadata}{ImageDatasetReference}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned experiments, e.g. a set of
\code{ExperimentIdentifier} or \code{Project} objects.}

\item{...}{Generic compatibility}
}
\description{
Experiment metadata can be fetched using \code{list_image_metadata()},
which accepts either a set of \code{Experiment}, \code{ExperimentIdentifier} or
\code{ImageDatasetReference} objects and returns all corresponding meta data as
\code{json_vec} of either \code{ExperimentImageMetadata}, in case experiment objects
were used or \code{ImageDatasetMetadata} objects, in case
\code{ImageDatasetReference} objects were used for calling
\code{list_image_metadata()}.
}
