% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_img_ref}
\alias{list_img_ref}
\alias{list_img_ref.NULL}
\alias{list_img_ref.WellPosition}
\title{List image references}
\usage{
list_img_ref(token, x, wells = NULL, channels, ...)

\method{list_img_ref}{NULL}(token, x, wells, channels, ...)

\method{list_img_ref}{WellPosition}(token, x, wells, channels, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit search for datasets/files with.}

\item{wells}{A (set of) \code{WellPosition} object(s) to limit the dataset
listing to.}

\item{channels}{A character vector with imaging channel names to limit the
dataset listing to.}

\item{...}{Generic compatibility.}
}
\description{
Used for double dispatching on the \code{list_datasets()} generic, list image
reference objects either for a specific (set of) \code{WellPosition} object(s)
or for the specified datasets in general.
}
\keyword{internal}
