% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{list_material}
\alias{list_material}
\alias{material_id}
\title{List materials}
\usage{
list_material(token, x)

material_id(code, type = "gene")
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{A (vector of) MaterialIdentifier object(s).}
}
\description{
List all materials available on the queried openBIS instance for a given
set of material id objects. Unfortunately there is no way for the available
API version to list material id objects. These have to be instantiated by
the user, using the function \code{material_id()}. A material id object is
defined by a material code and a material type. Current types are
}
\details{
\itemize{
\item compound
\item control
\item esirna
\item gene
\item mirna
\item mirna_inhibitor
\item mirna_mimic
\item pooled_sirna
\item sirna
}

The construction of material codes depends on material type, for example for
genes, the code is an ENTREZ id.
}
