% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_plate_datasets}
\alias{list_plate_datasets}
\title{Get data sets for a plate}
\usage{
list_plate_datasets(token, plate_id)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{plate_id}{Plate barcode.}
}
\value{
List, containing (among others), fields \"code\", and
\"dataSetTypeCode\".
}
\description{
Given a plate id (barcode), the corresponding sample object is
fetched using [get_plate_sample] and all available datasets for this sample
are queried.
}
