% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_plate_well_ref}
\alias{list_plate_well_ref}
\title{List plate/well references}
\usage{
list_plate_well_ref(token, material_id, experiment_id = NULL,
  include_datasets = FALSE)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{material_id}{Material id(s) corresponding to which wells are searched
for a collection of \code{MaterialIdentifierScreening} is expected.}

\item{experiment_id}{Additionally, the search can be limited to a single
experiment, specified either as \code{Experiment} or \code{ExperimentIdentifier}.}

\item{include_datasets}{Logical switch indicating whether to also return
the connected image and image analysis data sets.}
}
\description{
Given a login token, all wells available to the corresponding user on the
queried openBIS instance, which are associated with any of the given
material(s), are listed. If multiple material ids are searched for (e.g. a
\code{json_vec} of \code{MaterialIdentifierScreening}), an API call for each
object has to be made. The search can be limited to a single experiment,
specified either as \code{Experiment} or \code{ExperimentIdentifier}
}
