% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_plates}
\alias{list_plates}
\alias{list_plates.NULL}
\alias{list_plates.ExperimentIdentifier}
\alias{list_plates.Experiment}
\title{List plates}
\usage{
list_plates(token, x = NULL, ...)

\method{list_plates}{NULL}(token, x, ...)

\method{list_plates}{ExperimentIdentifier}(token, x, ...)

\method{list_plates}{Experiment}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned plates, e.g. a set of
\code{ExperimentIdentifier} or \code{Experiment} objects. The default is NULL, which
will fetch all available plates.}

\item{...}{Generic compatibility}
}
\description{
Given a login token, all plates available to the corresponding user on the
queried openBIS instance are listed for one or more experiment(s). If
multiple experiments are used for limiting the search (e.g. a \code{json_vec} of
experiments), an API call for each object has to be made. If no experiments
are specified, all available plates are returned.
}
