% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects(token, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{...}{Further arguments will be passed to \code{\link[=make_requests]{make_requests()}}.}
}
\value{
Depending on the number of resulting objects, either a
\code{\link{json_class}} (single object) or a \code{\link{json_vec}} (multiple objects), of
sub-type \code{Project} is returned.
}
\description{
A project forms one of the most basic entities in the organizational
hierarchy of openBIS. Each project consists of one or several experiments
and one or more projects are contained in each space, which is the topmost
structure used for grouping experiments. For InfectX, each project
corresponds to a separate pathogen. All registered projects visible to
the current user can be listed by calling the function \code{list_projects()}.
}
\section{openBIS}{

\itemize{
\item \Sexpr[results=rd]{infx::docs_link("gis", "listProjects")}
}
}

\examples{
\donttest{
  tok <- login_openbis()

  proj <- list_projects(tok)
  length(proj)
  get_field(proj, "code")

  logout_openbis(tok)
}

}
\seealso{
Other object listing functions: \code{\link{list_datasets}},
  \code{\link{list_experiments}},
  \code{\link{list_material}}, \code{\link{list_plates}},
  \code{\link{list_samples}}
}
\concept{object listing functions}
