% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_property_types}
\alias{list_property_types}
\title{List property types}
\usage{
list_property_types(token, with_relations = FALSE)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{with_relations}{Logical switch indicating whether relations should
be returned as well.}
}
\description{
Used primarily for searching, this function returns all property types
available throughout the queried openBis instance. As objects of several
types (\code{ControlledVocabularyPropertyType} and \code{PropertyType}) are returned
as property types by the API, the resulting object is a list with each
entry corresponding to a type and holding a set of object of the respective
type.
}
