% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{list_sample_types}
\alias{list_sample_types}
\title{List sample types}
\usage{
list_sample_types(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}
}
\value{
A a list of SampleType objects.
}
\description{
Given a login token, all available sample types are listed.
}
