% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_wells}
\alias{list_wells}
\alias{list_wells.PlateIdentifier}
\alias{list_wells.Plate}
\title{List wells}
\usage{
list_wells(token, x = NULL, ...)

\method{list_wells}{PlateIdentifier}(token, x, ...)

\method{list_wells}{Plate}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned wells, e.g. a set of
\code{PlateIdentifier} or \code{Plate} objects.}

\item{...}{Generic compatibility}
}
\description{
Given a login token, all wells available to the corresponding user on the
queried openBIS instance are listed for one or more plate(s). If multiple
plates are searched for (e.g. a \code{json_vec} of plates), an API call for each
object has to be made.
}
