% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_openbis}
\alias{login_openbis}
\alias{logout_openbis}
\alias{is_token_valid}
\title{Generate a login token}
\usage{
login_openbis(user, pwd, auto_disconnect = TRUE)

logout_openbis(token)

is_token_valid(token)
}
\arguments{
\item{user, pwd}{Login credentials for an openBis instance.}

\item{auto_disconnect}{Logical switch for automatically closing the
connection upon garbage collection of the token.}

\item{token}{Login token as created by \code{\link[=login_openbis]{login_openbis()}}.}
}
\value{
The login token to be used for further API interactions.

NULL (invisibly)

Scalar logical.
}
\description{
Create a login token for openBis API calls. Upon garbage
collection of the token, the user is logged out.

Using a token as created by \code{\link[=login_openbis]{login_openbis()}}, the
corresponding session is closed and the token is rendered invalid.

A token as created by \code{\link[=login_openbis]{login_openbis()}} is tested for its
validity.
}
\section{Functions}{
\itemize{
\item \code{login_openbis}: 

\item \code{logout_openbis}: 

\item \code{is_token_valid}: 
}}

\examples{
\donttest{
  tok <- login_openbis("rdgr2014", "IXPubReview", auto_disconnect = FALSE)
  is_token_valid(tok)
  logout_openbis(tok)
  is_token_valid(tok)
}

}
