% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbis.R
\name{login_openbis}
\alias{login_openbis}
\title{Generate a login token}
\usage{
login_openbis(user, pwd, auto_disconnect = TRUE, ...)
}
\arguments{
\item{user, pwd}{Login credentials for an openBis instance.}

\item{auto_disconnect}{Logical switch for automatically closing the
connection upon garbage collection of the token.}

\item{...}{Further arguments are passed to [query_openbis].}
}
\value{
The login token to be used for further API interactions.
}
\description{
Create a login token for openBis API calls. Upon garbage
collection of the token, the user is logged out.
}
