% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{logout_openbis}
\alias{logout_openbis}
\title{Logout from openBis}
\usage{
logout_openbis(token)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}
}
\value{
NULL (invisibly)
}
\description{
Using a token as created by [login_openbis], the corresponding
session is closed and the token is rendered invalid.
}
