% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{make_request}
\alias{make_request}
\title{Make a JSON-RPC request}
\usage{
make_request(url, method, params, version = "2.0", id = "1")
}
\arguments{
\item{url}{Url, the request is sent to.}

\item{method}{The method name}

\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The "@type" entries will be
generated from "json_class" attributes.}

\item{version}{JSON-RPC protocol version to be used.}

\item{id}{Id of the JSON-RPC request.}
}
\value{
A (nested) list holding the response from the JSON-RPC server
("@type" entries are converted to "json_class" attributes).
}
\description{
Issues a POST request to a JSON-RPC server. All "@type" fields
are converted to/from "json_class" attributes.
}
