% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{match_clause}
\alias{match_clause}
\alias{property_clause}
\alias{any_property_clause}
\alias{any_field_clause}
\alias{attribute_clause}
\alias{time_attribute_clause}
\alias{list_property_types}
\title{Create a match clause}
\usage{
match_clause(desired_value, clause_type = c("MatchClause",
  "PropertyMatchClause", "AnyPropertyMatchClause", "AnyFieldMatchClause",
  "AttributeMatchClause", "TimeAttributeMatchClause"), field_type = NULL,
  mode = "eq", ...)

property_clause(desired_value, property_code)

any_property_clause(desired_value)

any_field_clause(desired_value)

attribute_clause(desired_value, attribute = "code")

time_attribute_clause(desired_date = Sys.Date(), attribute = "registration",
  timezone = 0L, mode = "eq")

list_property_types(token, with_relations = FALSE, ...)
}
\arguments{
\item{desired_value}{The value used in the comparison.}

\item{clause_type}{The type of match clause.}

\item{field_type}{The type of filed used in the comparison.}

\item{mode}{The comparison mode, can be \code{eq} (==), \code{lte} (<=) or \code{gte}
(>=).}

\item{...}{Further arguments for creating the match clause.}

\item{property_code}{Code identifying a property to be used for the
comparison.}

\item{attribute}{Name of the attribute to be used for the comparison.}

\item{desired_date}{A date used for the comparison.}

\item{timezone}{A string identifying the timezone of the specified date,
examples include "+1", "-5", "0", etc.}

\item{token}{Login token as created by \code{login_openbis()}.}

\item{with_relations}{Logical switch indicating whether relations should
be returned as well.}
}
\description{
\code{SearchCriteria} objects, used for searching openBis, consist of one or
several match clauses. For more information on how to execute a search,
see \code{\link[=search_openbis]{search_openbis()}}. A match clause, broadly speaking, consist of a
given value, a field to which this value is compared to and a comparison
operator (e.g. equality).
}
\details{
In order to determine the possible values that can be supplied to
\code{property_clause()} as \code{property_code}s, \code{list_property_types()} can be
used. This function returns all property types available throughout the
queried openBis instance. As objects of several types
(\code{ControlledVocabularyPropertyType} and \code{PropertyType}) are returned as
property types by the API, the resulting object is a list with each entry
corresponding to a type and holding a set of object of the respective type.
}
