% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{print_json_class}
\alias{print_json_class}
\title{Helper function for printing JSON objects}
\usage{
print_json_class(x, depth, max_depth, layout = style())
}
\arguments{
\item{x}{The JSON object to print.}

\item{depth}{The current recursion depth.}

\item{max_depth}{The maximum recursion depth.}

\item{layout}{Characters for printing the tree structure and styles to be
applied to the different entities.}
}
\value{
A string that can be used for printing.
}
\description{
Inspired by the ast printing function of Hadley's [lobstr
package](https://github.com/hadley/lobstr) and also heavily borrowing code
from [there](https://git.io/vFMA5), this enables the recursive printing of
JSON objects.
}
