% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_openbis}
\alias{query_openbis}
\title{Helper querying openBis API}
\usage{
query_openbis(method, params, api = c("IGeneralInformationService",
  "IGeneralInformationChangingService", "IQueryApiServer",
  "IWebInformationService", "IDssServiceRpcGeneric", "IScreeningApiServer",
  "IDssServiceRpcScreening"), host = "https://infectx.biozentrum.unibas.ch")
}
\arguments{
\item{method}{The method name}

\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The "@type" entries will be
generated from "json_class" attributes.}

\item{api}{The api section the used method is part of. This is used to
construct the url the request is sent to.}

\item{host}{The base url, the request is sent to.}
}
\value{
A list/data.frame holding the response from openBis.
}
\description{
Issues a POST request to the JSON-RPC based openBis API v1,
using \link{make_request}. Documentation is available at
\code{https://wiki-bsse.ethz.ch/display/openBISDoc/openBIS+JSON+API}.
}
