% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{Read single cell data}
\usage{
read_data(file)
}
\arguments{
\item{file}{The filename of the file to be read.}
}
\value{
A vector holding all data. The attribute \code{length} can be used
to split the linearized data into bind corresponding to images.
}
\description{
Calling [R.matlab::readMat] single cell feature files in
Matlab format, as produced by CellProfiler are read and checked for the
correct structure. The Matlab file is expected to contain a nested list
where each node corresponds to an image and contains a list which is either
holding a single value or a vector of values.
}
