% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_delim}
\alias{read_delim}
\title{Read a delimited text file}
\usage{
read_delim(file, delim, col_types = readr::cols(.default =
  readr::col_character()), ..., col_spec = NULL)
}
\arguments{
\item{file, delim, col_types, ...}{All are passed to [readr::read_delim()].}

\item{col_spec}{Expects a named list with names corresponding to column
names of the input. Each node may contain a list with zero or more of
entries
\describe{
  \item{\code{name}}{The new column name in case it is to be renamed.}
  \item{\code{collector},\code{na},\code{locale}}{All passed to
    [readr::parse_vector()].}
  \item{\code{allow_na}}{Logical switch specifying whether NA is allowed.}
  \item{\code{processor}}{A function that is applied to the column after
    reading/parsing for further transformations.}
}}
}
\value{
A tibble holding all read meta data or a subset thereof, depending
on the column specification.
}
\description{
Read tabular data structured as delimited text file, using
[readr::read_delim()]. All arguments except \code{col_spec} are passed to
[readr::read_delim()], while \code{col_spec} provides some additional
options for controlling input parsing.
}
