% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_full_meta}
\alias{read_full_meta}
\title{Read full meta data}
\usage{
read_full_meta(dat, col_types = list(.default = readr::col_character()), ...)
}
\arguments{
\item{dat}{The data (raw vector) to be read.}

\item{col_types}{A column specification to be passed to [readr::read_delim].}

\item{...}{All further arguments are passed to [readr::read_delim]. If an
argument should behave differently for each table, a named list is expected
with names
\enumerate{
  \item \code{\"well_annotation\"}
  \item \code{\"sequence_information_sirna\"}
  \item \code{\"sequence_information_mirna\"}
  \item \code{\"sequence_information_esirna\"}
  \item \code{\"sequence_information_compound\"}
  \item \code{\"pool_contained_compound_lookup\"}
}
for this argument in order to match to the respective table.}
}
\value{
List of tibbles holding all read meta data or a subset thereof,
depending on the column specification.
}
\description{
Read full meta data downloaded from openBis using
[readr::read_delim].
}
