% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_meta}
\alias{read_meta}
\title{Read meta data}
\usage{
read_meta(file, col_types = readr::cols(.default = readr::col_character()),
  ...)
}
\arguments{
\item{file}{The filename of the file to be read.}

\item{col_types}{A column specification to be passed to [readr::read_delim].}

\item{...}{All further arguments are passed to [readr::read_delim].}
}
\value{
A tibble holding all read meta data or a subset thereof, depending
on the column specification.
}
\description{
Read meta data downloaded from openBis using
[readr::read_delim]. The delimiting character is either ";" for public meta
data or "\\t" for dumped compound database files.
}
