% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-utils.R
\name{remove_null}
\alias{remove_null}
\title{Remove NULL entries}
\usage{
remove_null(x)
}
\arguments{
\item{x}{Object to process.}
}
\description{
Recursively remove all NULL fields from a nested list structure while
preserving \code{json_class} and \code{json_vec} class attributes.
}
\examples{
tmp <- json_class(a = json_class(b = "c", d = NULL, class = "foo"),
                  e = json_class(f = "g", class = "bar"),
                  h = NULL,
                  class = "foobar")
print(tmp, 2)
print(remove_null(tmp), 2)

}
