% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_request}
\alias{make_request}
\alias{query_openbis}
\title{Make a JSON-RPC request}
\usage{
make_request(url, method, params, version = "2.0", id = "1")

query_openbis(method, params, api = c("IGeneralInformationService",
  "IGeneralInformationChangingService", "IQueryApiServer",
  "IWebInformationService", "IDssServiceRpcGeneric", "IScreeningApiServer",
  "IDssServiceRpcScreening"), host = "https://infectx.biozentrum.unibas.ch")
}
\arguments{
\item{url}{Destination url, the request is sent to.}

\item{method}{The API method name.}

\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The \code{@type} entries will be
generated from \code{json_class} attributes.}

\item{version}{JSON-RPC protocol version to be used (defaults to \code{"2.0"}.}

\item{id}{An identifier for the JSON-RPC request (defaults to \code{"1"}). Can be
ignored, as only single JSON-RPC requests are issued per HTTP request.}

\item{api, }{host Strings used to construct the destination url.}
}
\value{
A (nested) list holding the response from the JSON-RPC server
(\code{@type} entries are converted to \code{json_class} attributes).
}
\description{
Issues a POST request to a JSON-RPC server. All \code{@type} fields are
converted to/from \code{json_class} attributes, using \code{\link[=as_json_list]{as_json_list()}} and
\code{\link[=as_json_class]{as_json_class()}}. The helper function \code{query_openbis()} wraps
\code{make_request()} and constructs the url the request is sent to.
}
\examples{
\donttest{
  tok <- login_openbis("rdgr2014", "IXPubReview")
  projects <- query_openbis("listProjects", tok)
  print(projects[[1]])
}

}
