% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{style}
\alias{style}
\title{Styles for printing JSON objects}
\usage{
style(fancy = TRUE)
}
\arguments{
\item{fancy}{Logical switch to enable font styles, colors and UTF box
characters.}
}
\value{
A list holding the tree characters and the entity styling.
}
\description{
Characters for printing the tree structure and styles to be
applied to the different entities in [print_json_class].
}
