% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_code}
\alias{dataset_code}
\title{Extract dataset code}
\usage{
dataset_code(x, ...)
}
\arguments{
\item{x}{Dataset object(s).}

\item{...}{Generic compatibility.}
}
\description{
Given a (set of) dataset object(s), for each one extract the dataset code
and return a character vector of codes.
}
