% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{do_download}
\alias{do_download}
\title{Download files}
\usage{
do_download(token, data_id, files = NULL, rep = 2)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{data_id}{Data set code.}

\item{files}{Files objects as produced by [list_files] describing the files
to be downloaded. If NULL, all available files of this data_id will be
fetched (as long as there are less than 10).}

\item{rep}{The number of times failed downloads are repeated.}
}
\value{
A list of raw vectors holding the downloaded data.
}
\description{
Given a data set code and a small set (<= 10) of file paths
(as returned from [list_files]), a download link is generated for each file
by [get_download]. The downloads are preformed asynchronously.
}
\section{TODO}{
 use file checksums instead of length for checking file
integrity; see https://stackoverflow.com/q/21558777, can't rawToChar, might
have to write to disk?

 rename to async_download and also provide a single_download
option
}

