% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_default}
\alias{error_default}
\title{Helper for S3 default error}
\usage{
error_default(x, prefix = "")
}
\arguments{
\item{x}{The object on which the dispatch is done.}

\item{prefix}{Optional string that is inserted before the error message.}
}
\description{
Whenever the default function of an S3 generic is a \code{not implemented} error,
this function can be used. It is a wrapper around \code{\link[base:stop]{base::stop()}}, with a
custom error message.
}
\keyword{internal}
