% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{fetch_files}
\alias{fetch_files}
\alias{fetch_files.FileInfoDssDTO}
\alias{fetch_files_serial}
\alias{fetch_files_parallel}
\title{Fetch files}
\usage{
fetch_files(token, x, ...)

\method{fetch_files}{FileInfoDssDTO}(token, x, data_sets, n_con = 5L, ...)

fetch_files_serial(urls, n_try = 2L, file_sizes = NULL, done = identity,
  ...)

fetch_files_parallel(urls, n_try = 2L, file_sizes = NULL, done = identity,
  n_con = 5L, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to specify which files to download.}

\item{...}{Generic compatibility.}

\item{data_sets}{Either a single dataset object (anything that has a
\code{dataset_code()} method) or a set of objects of the same length as \code{x}.}

\item{n_con}{The number of simultaneous connections.}

\item{urls}{Either a caracter vector or a list of calls that each yields an
url when \code{eval}d.}

\item{n_try}{The number of tries for each url.}

\item{file_sizes}{A vector of expected file sizes or NULL.}

\item{done}{A function with a single argument which is applied to each
downloaded file.}
}
\description{
The function \code{fetch_files()} downloads files specified by a set of
\code{FileInfoDssDTO} objects alongside one or several dataset objects.
}
