% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_files}
\alias{fetch_files}
\title{Split downloads}
\usage{
fetch_files(token, files, async = 5L)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{files}{Files objects as produced by [list_files] describing the files
to be downloaded. If NULL, all available files of this data_id will be
fetched (as long as there are less than 10).}

\item{async}{The bucket size for simultaneous downloads.}
}
\value{
A list of raw vectors holding the downloaded data.
}
\description{
A set of files is split into buckets and downloaded
asynchronously and the progress is reported using a progress bar. The
splitting is necessary as [do_download] downloads all files simultaneously.
}
\section{TODO}{
 entry point to all downloads, choose single or async (both
not exported) and use this in plate/meta fetching.
}

