% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_meta}
\alias{fetch_meta}
\title{Fetch InfectX meta data}
\usage{
fetch_meta(token, type = c("full", "public"), ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{type}{A switch for the type of meta data to be downloaded.}

\item{...}{Further arguments are passed to [query_openbis].}
}
\value{
A list of downloaded files (raw).
}
\description{
This function essentially provides default values for
downloading meta data from openBis. Two formats of meta data are currently
available: dumps of SQLite databased holding the entire set of experimental
meta data, as well as a CSV sheet, containing only the published subset.
}
