% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fetch_plate}
\alias{fetch_plate}
\title{Fetch single cell data}
\usage{
fetch_plate(plate_name, ...)
}
\arguments{
\item{plate_name}{The plate barcode of interest.}

\item{...}{Arguments passed to [fetch_data()]}
}
\value{
A vector of downloaded filenames.
}
\description{
Calling [fetch_openbis()], this function downloads single cell
datasets corresponding to a plate barcode and filtered by an optional file
name regular expression.
}
