% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_plate}
\alias{fetch_plate}
\title{Fetch single cell data}
\usage{
fetch_plate(token, plate_id, file_regex, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{plate_id}{Plate barcode.}

\item{file_regex}{The plate barcode of interest.}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
A list of downloaded file data (raw).
}
\description{
Download single cell datasets corresponding to a plate barcode
and filtered by a regular expression applied to file names.
}
