% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_download}
\alias{get_download}
\title{Get download link for file}
\usage{
get_download(token, data_id, file, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{data_id}{Data set code.}

\item{file}{File path for which the link is generated.}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
Url linking to file (expires immediately).
}
\description{
Given a data set code and a file path, a download link is
generated, which has to be consumed immediately.
}
