% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbis.R
\name{get_plate_sample}
\alias{get_plate_sample}
\title{Get sample object of plate}
\usage{
get_plate_sample(token, plate_id, space_code = NULL, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{plate_id}{Plate barcode.}

\item{space_code}{The space code of the plate; it NULL, it is determined
automatically.}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
List, containing (among others), entries \"id\", \"permId\",
\"identifier\", \"properties\", \"retrievedFetchOptions\".
}
\description{
Given a plate id (barcode), the corresponding plate space is
determined using [list_plates] and the sample object representing the given
plate is queried.
}
