% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infx.R
\docType{package}
\name{infx}
\alias{infx}
\alias{infx-package}
\title{API access to the InfectX data repository}
\description{
The JSON-RPC based openBIS API can be conveniently accesses form R using
functionality provided with this package. The intended application is data
retrieval from the openBIS instance hosted by the InfectX high throughput
screening project. Therefore some aspects of the API geared more towards
data curation are currently not supported. However the basic infrastructure
for creating and issuing a request, as well as processing the response,
is exposed and missing functionality therefore could easily be added.
}
\details{
Type information of JSON objects returned from the API is preserved as S3
class attribute and all JSON list structures additionally inherit from the
S3 class \code{json_class}. As such, a data set object retrieved form openBIS,
for example, will have two class attributes: \code{DataSet} and \code{json_class}.
Sets of \code{json_class} objects that are of the same sub-type can be
represented as \code{json_vec} objects of that sub-type. Several data set
objects therefore are assembled as list structure with S3 classes \code{DataSet}
and \code{json_vec}, where every entry in turn is an S3 object with types
\code{DataSet} and \code{json_class}. This approach was choses in order to not only
have available generic function dispatch on individual \code{json_class}
objects, but also on sets (or \emph{vectors}) of \code{json_class} objects.

This documentation makes a distinction between objects in openBIS that exist
mainly for the purpose of organizing and grouping data and objects that
represent actual data resources. The most basic object in the organizational
hierarchy is that of a \code{Project}. Several \code{Experiment} objects may be
associated with a \code{Project} and \code{Sample} objects live in experiments. Given
the HTS-based context of InfectX data, samples typically represent
microtiter plates and individual plate wells. \code{Material} objects describe
agents applied to samples. Many of the InfectX screens are RNA interference-
based and therefore materials may be for example siRNA oligos or targeted
genes. Finally, samples are associated with \code{DataSet} objects that stand for
experimental measurements or data derived thereof.

Any type of data resource available in openBIS can be accessed as files
belonging to data sets. Due to the image-based nature of InfectX screens,
raw experimental data comes in the form of fluorescence microscopy imagery
which consequently constitutes the most basic form of data resource
available. It is therefore no surprise that image data receives special
treatment, allowing for more fine grained access and functionality that
helps with finding specific sub-sets of images. A further data resource
that comes with features similar to those of image data is termed feature
vector data sets. This is mostly tabular data with a single value
corresponding to an imaging site. This is typically used for image
acquisition meta data, summarized image analysis or quality control results.
}
