% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{is_token_valid}
\alias{is_token_valid}
\title{Check validity of token}
\usage{
is_token_valid(token, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
Scalar logical.
}
\description{
A token as created by [login_openbis] is tested for its
validity.
}
