% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-utils.R
\name{has_fields}
\alias{has_fields}
\alias{has_fields.json_class}
\alias{has_fields.json_vec}
\title{Check JSON objects}
\usage{
has_fields(x, fields, ...)

\method{has_fields}{json_class}(x, fields, ...)

\method{has_fields}{json_vec}(x, fields, ...)
}
\arguments{
\item{x}{Object to test.}

\item{fields}{Character vector of nonzero length, holding the field names
for which to check.}

\item{...}{Generic compatibility.}
}
\description{
Test whether a single \code{json_class} object contains all of the specified
fields or whether each \code{json_class} object contained in a \code{json_vec} object
passes this test.
}
\examples{
obj_1 <- json_class(a = 1, b = 2, class = "foo")
obj_2 <- json_class(a = 3, b = 4, class = "foo")
obj_3 <- json_class(a = 3, c = 4, class = "foo")

has_fields(obj_1, "a")
has_fields(obj_1, c("a", "b"))

has_fields(c(obj_1, obj_2), "a")
has_fields(c(obj_1, obj_3), "a")
has_fields(c(obj_1, obj_3), c("a", "b"))

}
