% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-class.R
\name{json_class}
\alias{json_class}
\alias{new_json_class}
\alias{as_json_class}
\alias{as.json_class}
\alias{as_json_class.json_class}
\alias{as_json_class.list}
\alias{as_json_class.json_vec}
\alias{as_json_class.default}
\alias{rm_json_class}
\alias{as.list.json_class}
\alias{is_json_class}
\alias{is.json_class}
\alias{check_json_class}
\alias{has_json_subclass}
\alias{get_json_subclass}
\alias{print.json_class}
\title{JSON class objects}
\usage{
json_class(..., class)

new_json_class(x, class = NULL)

as_json_class(x, ...)

as.json_class(x, ...)

\method{as_json_class}{json_class}(x, ...)

\method{as_json_class}{list}(x, ...)

\method{as_json_class}{json_vec}(x, ...)

\method{as_json_class}{default}(x, force = FALSE, ...)

rm_json_class(x, recursive = TRUE, restore_type = TRUE)

\method{as.list}{json_class}(x, keep_asis = TRUE, recursive = !keep_asis,
  restore_type = !keep_asis, ...)

is_json_class(x)

is.json_class(x)

check_json_class(x, recursive = TRUE)

has_json_subclass(x, class)

get_json_subclass(x)

\method{print}{json_class}(x, depth = 1L, width = getOption("width"),
  length = 100L, fancy = TRUE, ...)
}
\arguments{
\item{...}{Generic compatibility.}

\item{class}{JSON sub-class name.}

\item{x}{Object to process.}

\item{force}{Suppress error when casting an object to \code{json_class} that
cannot be converted.}

\item{recursive}{Recursively apply the function.}

\item{restore_type}{When removing the \code{json_class} information from an
object, whether to preserve the subclass attribute as \code{@type} filed.}

\item{keep_asis}{Used in \code{as.list()}, if \code{TRUE}, the \code{json_class} object
is returned as-is, if \code{FALSE}, class attributes may be dropped/written to
the list structure into the \code{@type} field.}

\item{depth}{The maximum recursion depth for printing.}

\item{width}{Number of columns to maximally print.}

\item{length}{Number of lines to maximally print.}

\item{fancy}{Logical switch to enable font styles, colors and UTF box
characters for printing.}
}
\description{
To communicate object type information via JSON to the Jackson-powered
openBis interface, the \code{@type} field is used. Data received from openBis is
recursively stripped of \code{@type} fields and the type information is saved as
class attribute. Such objects also have the class \code{json_class} added. The
function \code{as_json_class()} (or its alias \code{as.json_class()}) powers this
recursive conversion of a list with filed \code{@type} into a \code{json_class}
object. The constructor \code{json_class()} is available for non-recursive
instantiation of \code{json_class} objects.
}
\details{
The action of \code{as_json_class()} is reversed by \code{rm_json_class()}. This
removes both the \code{json_class} class attribute and the JSON class attribute
itself, which is subsequently written to a \code{@type} filed. This preserving
of type information can be disabled, by setting the argument \code{restore_type}
to \code{FALSE}. Furthermore, the action can be applied recursively with the
argument \code{recursive}. The function \code{as.list()} can also be used to perform
the above actions, but with default arguments, it does nothing, as
functions such as \code{base::sapply()} and \code{base::lapply()}, call \code{as.list()}.

The functions \code{is_json_class()} and \code{has_json_subclass()} test whether an
object is a JSON class object. The former tests whether an object is a
proper \code{json_class} object, meaning that:
\itemize{
\item it is a list
\item it inherits \code{json_class}
\item the last class attribute is \code{json_class}
\item apart from \code{json_class} there exists at least one more class attribute
The latter function tests whether an object has a specific JSON class
attached. In order to recursively test a \code{json_class} object for being
properly formed, the function \code{check_json_class()} can be used.
}

JSON class objects have custom sub-setting and printing functions available.
Sub-setting of JSON objects that preserve class and \code{json_class}
attributes. This is useful when objects are created from openBIS results
which are subsequently used in further queries, but the constructors they
are passed to require only a subset of the fetched fields. Printing is
inspired by (and heavily borrows code from) the ast printing function of
Hadley's \href{https://github.com/hadley/lobstr}{lobstr package}.
}
\examples{
lst <- list(`@type` = "foo", "a", "b")
cls <- as_json_class(lst)

print(cls)
is_json_class(cls)
get_json_subclass(cls)

identical(rm_json_class(cls), lst)

}
