% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-class.R
\name{is_json_class}
\alias{is_json_class}
\alias{is.json_class}
\alias{check_json_class}
\title{Validate JSON class objects}
\usage{
is_json_class(x)

is.json_class(x)

check_json_class(x, recursive = TRUE)
}
\arguments{
\item{x}{Object to process.}

\item{recursive}{Recursively apply the function.}
}
\description{
The functions \code{is_json_class()} tests whether an object is a proper JSON
class object, meaning that:
\itemize{
\item it is a list
\item it inherits \code{json_class}
\item the last class attribute is \code{json_class}
\item apart from \code{json_class} there exists at least one more class attribute
}
}
\details{
In order to recursively test a \code{json_class} object for being properly
formed, the function \code{check_json_class()} can be used. This recurses through
a list structure and whenever an object inherits from \code{json_class} it is
tested with \code{is_json_class()}.
}
\examples{
lst <- list(`@type` = "foo", "a", "b")
cls <- as_json_class(lst)

print(cls)
is_json_class(cls)
get_subclass(cls)

identical(rm_json_class(cls), lst)

}
