% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_datasets}
\alias{list_datasets}
\alias{list_datasets.Sample}
\alias{list_datasets.Experiment}
\alias{list_datasets.character}
\title{List datasets}
\usage{
list_datasets(token, x, ...)

\method{list_datasets}{Sample}(token, x, include = c(NA, "children",
  "parents", "all"), ...)

\method{list_datasets}{Experiment}(token, x, include = c(NA, "children",
  "parents", "all"), ...)

\method{list_datasets}{character}(token, x, include = c(NA, "children",
  "parents", "all"), ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit search for datasets with.}

\item{...}{Generic compatibility.}

\item{include}{Whether to include parent/child datasets as well.}
}
\description{
Given a login token, all available datasets are listed for the given
experiment(s), sample(s) or dataset code(s). Additionally it can be
specified whether parent or child datasets are to be included as well.
}
\section{TODO}{
 The API function \code{listDataSetsForSample()} has a parameter
\code{areOnlyDirectlyConnectedIncluded}, which is currently fixed to \code{TRUE}. The
documentation contains the following explanation:

If true, only data sets that are directly connected to the sample are
included, otherwise data sets of child samples are included as well.

This does however not seem to correspond to including child datasets in the
API call to \code{listDataSets()} via its \code{connectionsToGet} argument. As long
as it is not entirely clear how the inclusion of child/parent datasets
differs from setting \code{areOnlyDirectlyConnectedIncluded} to \code{FALSE}, this
option is not exposed to the user.
}

