% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{list_datasets}
\alias{list_datasets}
\title{Get data sets for a plate}
\usage{
list_datasets(token, plate_id, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{plate_id}{Plate barcode.}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
List/data.frame, containing (among others), columns \"code\",
\"dataSetTypeCode\".
}
\description{
Given a plate id (barcode), the corresponding sample object is
fetched using [get_plate_sample] and all available datasets for this sample
are queried.
}
