% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_datasets}
\alias{list_datasets}
\alias{list_datasets.Sample}
\alias{list_datasets.Experiment}
\alias{list_datasets.character}
\alias{list_dataset_types}
\alias{list_dataset_ids}
\alias{list_files}
\alias{list_files.character}
\title{List datasets}
\usage{
list_datasets(token, x, ...)

\method{list_datasets}{Sample}(token, x, include = c(NA, "children",
  "parents", "all"), ...)

\method{list_datasets}{Experiment}(token, x, include = c(NA, "children",
  "parents", "all"), ...)

\method{list_datasets}{character}(token, x, include = c(NA, "children",
  "parents", "all"), ...)

list_dataset_types(token)

list_dataset_ids(token, codes)

list_files(token, x, ...)

\method{list_files}{character}(token, x, path = "", recursive = TRUE)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit search for datasets/files with.}

\item{...}{Generic compatibility.}

\item{include}{Whether to include parent/child datasets as well.}

\item{codes}{Character vector of one or more dataset codes.}

\item{path}{A (vector of) file path(s) to be searched within a dataset.}

\item{recursive}{A (vector of) logicals, indicating whether to list files
recursively.}
}
\description{
Given a login token, all available datasets are listed for the given
experiment(s), sample(s) or dataset code(s). Additionally it can be
specified whether parent or child datasets are to be included as well.
\code{list_dataset_types()} lists all available data set types on the given
openBIS instance and \code{list_dataset_id()} returns a list of dataset id
objects corresponding to the supplied character vector of one or more
dataset codes.
}
\details{
The function \code{list_files()} lists files associated with one or more
dataset(s), which can be specified with codes or \code{DataSet}/
\code{DatasetIdentifier} objects.
}
\section{TODO}{
 The API function \code{listDataSetsForSample()} has a parameter
\code{areOnlyDirectlyConnectedIncluded}, which is currently fixed to \code{TRUE}. The
documentation contains the following explanation:

If true, only data sets that are directly connected to the sample are
included, otherwise data sets of child samples are included as well.

This does however not seem to correspond to including child datasets in the
API call to \code{listDataSets()} via its \code{connectionsToGet} argument. As long
as it is not entirely clear how the inclusion of child/parent datasets
differs from setting \code{areOnlyDirectlyConnectedIncluded} to \code{FALSE}, this
option is not exposed to the user.
}

