% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_datastores}
\alias{list_datastores}
\title{List data store servers}
\usage{
list_datastores(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}
}
\description{
List all data store servers registered this openBIS server instance.
}
