% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{list_experiment_ids}
\alias{list_experiment_ids}
\title{List experiment ids}
\usage{
list_experiment_ids(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}
}
\value{
A a list of ExperimentIdentifier objects.
}
\description{
Given a login token, all available experiment ids are listed.
}
