% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_experiment_ids}
\alias{list_experiment_ids}
\title{List experiment types}
\usage{
list_experiment_ids(token)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}
}
\value{
List/data.frame, containing (among others), columns \"plateCode\"
and \"spaceCodeOrNull\".
}
\description{
For a login token, list all available experiment types.
}
