% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{list_experiment_types}
\alias{list_experiment_types}
\title{List experiment types}
\usage{
list_experiment_types(token, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
List/data.frame, containing (among others), columns \"plateCode\"
and \"spaceCodeOrNull\".
}
\description{
For a login token, list all available experiment types.
}
