% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{list_experiments}
\alias{list_experiments}
\title{List experiments}
\usage{
list_experiments(x, ...)
}
\arguments{
\item{x}{Object to limit the number of returned experiments, e.g. a set of
\code{ExperimentIdentifier} objects.}
}
\value{
A a list of ExperimentIdentifier objects.
}
\description{
Given a login token and a set of experiment ids, all available experiments
are listed.
}
