% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_experiments}
\alias{list_experiments}
\title{List plates}
\usage{
list_experiments(token, projects = NULL, exp_type = NULL, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{projects}{List holding Project objects for which experiments are to be
listed. If NULL, all experiments are returned.}

\item{exp_type}{Character vector, specifying the desired experiment type.
If NULL, all experiments are returned.}

\item{...}{Further arguments are passed to [query_openbis].}
}
\value{
List/data.frame, containing (among others), columns \"plateCode\"
and \"spaceCodeOrNull\".
}
\description{
For a login token, list all available plates.
}
