% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{list_files}
\alias{list_files}
\alias{list_files.character}
\alias{list_files.DataSet}
\alias{list_files.DatasetIdentifier}
\alias{list_files.DataSetFileDTO}
\title{List files}
\usage{
list_files(token, x, ...)

\method{list_files}{character}(token, x, path = "", recursive = TRUE, ...)

\method{list_files}{DataSet}(token, x, path = "", recursive = TRUE, ...)

\method{list_files}{DatasetIdentifier}(token, x, path = "",
  recursive = TRUE, ...)

\method{list_files}{DataSetFileDTO}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit search for datasets/files with.}

\item{...}{Generic compatibility.}

\item{path}{A (vector of) file path(s) to be searched within a dataset.}

\item{recursive}{A (vector of) logicals, indicating whether to list files
recursively.}
}
\description{
The function \code{list_files()} lists files associated with one or more
dataset(s), which can be specified with codes or \code{DataSet}/
\code{DatasetIdentifier} objects.
}
