% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_files}
\alias{list_files}
\title{Get files for a data set}
\usage{
list_files(token, data_id, folder = "original")
}
\arguments{
\item{token}{Login token as created by \link{login_openbis}.}

\item{data_id}{Data set code.}

\item{folder}{Folder to which the file search is restricted.}
}
\value{
List/data.frame, containing (among others), columns
"pathInDataSet", "pathInListing".
}
\description{
Given a data set code, the corresponding files are queried.
}
