% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{list_image_metadata}
\alias{list_image_metadata}
\title{List image metadata}
\usage{
list_image_metadata(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned experiments, e.g. a set of
\code{ExperimentIdentifier} or \code{Project} objects.}

\item{...}{Generic compatibility}
}
\description{
Experiment metadata can be fetched using \code{list_image_metadata()},
which accepts either a set of \code{Experiment} or \code{ExperimentIdentifier}
objects and returns all corresponding metadata as \code{json_vec} of
\code{ExperimentImageMetadata} objects.
}
