% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{list_img_datasets}
\alias{list_img_datasets}
\title{List image datasets}
\usage{
list_img_datasets(token, plate_id, type = c("raw", "segmentation"),
  most_recent = TRUE)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{plate_id}{Either a character vector holding plate barcodes or a list
of/single PlateIdentifier object(s).}

\item{type}{A switch to choose from raw or segmentation image datasets.}

\item{most_recent}{Logical switch, if TRUE only the most recent dataset is
returned.}
}
\value{
A list/single object of type ImageDatasetReference.
}
\description{
For a plate, which can be specified as a (vector of)
string(s), for which the PlateIdentifier objects then are constructed using
[create_plate_id], or for a (list of) PlateIdentifier object(s), list image
datasets.
}
