% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_plates}
\alias{list_plates}
\alias{list_plates.NULL}
\alias{list_plates.ExperimentIdentifier}
\alias{list_plates.Experiment}
\alias{as_plateid}
\alias{as_plateid.Plate}
\alias{as_plateid.Sample}
\alias{list_wells}
\alias{list_wells.PlateIdentifier}
\alias{list_wells.Plate}
\alias{list_plate_well_ref}
\alias{list_plate_metadata}
\alias{list_plate_metadata.PlateIdentifier}
\alias{list_plate_metadata.Plate}
\title{List wells/plates}
\usage{
list_plates(token, x = NULL, ...)

\method{list_plates}{NULL}(token, x, ...)

\method{list_plates}{ExperimentIdentifier}(token, x, ...)

\method{list_plates}{Experiment}(token, x, ...)

as_plateid(x, ...)

\method{as_plateid}{Plate}(x, ...)

\method{as_plateid}{Sample}(x, ...)

list_wells(token, x, ...)

\method{list_wells}{PlateIdentifier}(token, x, ...)

\method{list_wells}{Plate}(token, x, ...)

list_plate_well_ref(token, material_id, experiment = NULL,
  include_datasets = FALSE)

list_plate_metadata(token, x, ...)

\method{list_plate_metadata}{PlateIdentifier}(token, x, ...)

\method{list_plate_metadata}{Plate}(token, x, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned wells or plates.}

\item{...}{Generic compatibility}

\item{material_id}{Material id(s) corresponding to which wells are searched
for a collection of \code{MaterialIdentifierScreening} is expected.}

\item{experiment}{Additionally, the search can be limited to a single
experiment, specified either as \code{Experiment} or \code{ExperimentIdentifier}.}

\item{include_datasets}{Logical switch indicating whether to also return
the connected image and image analysis data sets.}
}
\description{
Using \code{list_plates()}, all plates available to the user (corresponding to
the given login token) on the queried openBIS instance are listed for one
or more experiment(s). If multiple experiments are used for limiting the
search (e.g. a \code{json_vec} of experiments), an API call for each object has
to be made. If no experiments are specified, all available plates are
returned.
}
\details{
Wells can be listed with \code{list_wells()}, which given a login token, lists
all wells available to the corresponding user on the queried openBIS
instance for one or more plate(s). If multiple plates are searched for
(e.g. a \code{json_vec} of plates), an API call for each object has to be made.

A convenience function that allows for the conversion of plate objects into
plate id objects is available as \code{as_plateid()}. This function does
not incur an API call and can act on both a single plate object or a vector
of plate objects (passed as a \code{json_vec} object) and will return a
\code{json_vec} object of type \code{PlateIdentifier}.

\code{list_plate_well_ref()} can list well references (as
\code{PlateWellReferenceWithDatasets} objects) corresponding to materials. This
can be used to find all wells that contain a certain material (for example
a gene knockdown or a specific compound used for a knockdown). If multiple
material ids are searched for (e.g. a \code{json_vec} of
\code{MaterialIdentifierScreening}), an API call for each object has to be made.
The search can be limited to a single experiment,  specified either as
\code{Experiment} or \code{ExperimentIdentifier}. As a further argument, the switch
\code{include_datasets} specifies whether the connected image and image analysis
data sets should be returned as well.

All metadata corresponding to a plate (specified as a set of \code{Plate} or
\code{PlateIdentifier} objects) can be fetched using \code{list_plate_metadata()}.
The returned \code{PlateMetadata} objects also contain a list of corresponding
\code{WellMetadata} objects.
}
\section{openBIS}{

\itemize{
\item \Sexpr{infx::docs_link("sas", "listPlates")}
\item \Sexpr{infx::docs_link("sas", "listPlateWells")}
\item \Sexpr{infx::docs_link("sas", "getPlateMetadataList")}
}
}

