% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_plates}
\alias{list_plates}
\title{List plates}
\usage{
list_plates(token, exp_id = NULL)
}
\arguments{
\item{token}{Login token as created by \link{login_openbis}.}

\item{exp_id}{A single ExperimentIdentifier object to limit the plate
search to.}
}
\value{
List/data.frame, containing (among others), columns "plateCode"
and "spaceCodeOrNull".
}
\description{
For a login token, list all available plates.
}
