% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbis.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects(token, ...)
}
\arguments{
\item{token}{Login token as created by [login_openbis].}

\item{...}{Further arguments are passed to [do_openbis].}
}
\value{
List/data.frame, containing (among others), columns \"plateCode\"
and \"spaceCodeOrNull\".
}
\description{
For a login token, list all available projects.
}
