% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{list_projects}
\alias{list_projects}
\title{List projects}
\usage{
list_projects(token, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{...}{Further arguments will be passed to \code{\link[=make_requests]{make_requests()}}.}
}
\description{
List all projects available on the queried openBIS instance. A project forms
one of the most basic entities in the organizational hierarchy of openBIS.
One or more projects are contained in each space (the topmost
organizational entity) and each project consists of one or several
experiments.
}
\section{openBIS}{

\itemize{
\item \Sexpr{infx::docs_link("gis", "listProjects")}
}
}

