% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{list_samples}
\alias{list_samples}
\alias{list_samples.ExperimentIdentifier}
\alias{list_samples.Experiment}
\alias{list_samples.PlateIdentifier}
\alias{list_samples.Plate}
\alias{list_sample_types}
\title{List samples}
\usage{
list_samples(token, x, ...)

\method{list_samples}{ExperimentIdentifier}(token, x, ...)

\method{list_samples}{Experiment}(token, x, ...)

\method{list_samples}{PlateIdentifier}(token, x, ...)

\method{list_samples}{Plate}(token, x, ...)

list_sample_types(token)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{x}{Object to limit the number of returned samples, e.g. a set of
\code{ExperimentIdentifier} or \code{Experiment} objects.}

\item{...}{Generic compatibility}
}
\description{
Given a login token, all samples available to the corresponding user on the
queried openBIS instance are listed for an experiment, plate or well. If
multiple objects are used for limiting the search (e.g. a \code{json_vec} of
wells), an API call for each object has to be made. Furthermore, all
available sample types can be listed using \code{list_sample_types()}.
}
\section{TODO}{

Add class specific functions to generic \code{list_samples()} for
classes \code{PlateIdentifier} and \code{WellIdentifier} (as well as \code{Plate} and
\code{Well}).
}

