% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{list_datastores}
\alias{list_datastores}
\alias{list_datastore_urls}
\alias{list_download_urls}
\alias{list_download_urls.character}
\alias{list_download_urls.DataSet}
\alias{list_download_urls.DatasetIdentifier}
\alias{list_download_urls.DataSetFileDTO}
\title{List data store servers}
\usage{
list_datastores(token)

list_datastore_urls(token, data_set = NULL)

list_download_urls(token, x, ...)

\method{list_download_urls}{character}(token, x, path, timeout = NA, ...)

\method{list_download_urls}{DataSet}(token, x, path, timeout = NA, ...)

\method{list_download_urls}{DatasetIdentifier}(token, x, path, timeout = NA,
  ...)

\method{list_download_urls}{DataSetFileDTO}(token, x, timeout = NA, ...)
}
\arguments{
\item{token}{Login token as created by \code{login_openbis()}.}

\item{data_set}{Set of datasets (specified as dataset codes, \code{DataSet} or
\code{DatasetIdentifier} objects) or NULL.}

\item{x}{Object representing a (set of) dataset(s), e.g. a vector of dataset
codes, or a set of \code{DataSet}s or \code{DatasetIdentifier}s.}

\item{...}{Generic compatibility.}

\item{path}{A character vector of file paths within datasets.}

\item{timeout}{Time-span (in seconds) for which the file download link
should be valid.}
}
\description{
\code{list_datastores()} lists all data store servers registered this openBIS
server instance and \code{list_datastore_urls()} returns the base urls of the
data store servers that host the supplied datasets (or that of the default
data store server in case no data set is supplied). Download urls of files
can be listed using the function \code{list_download_urls()}.
}
