% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_openbis}
\alias{login_openbis}
\alias{logout_openbis}
\alias{is_token_valid}
\title{Generate/check/destroy a login token}
\usage{
login_openbis(user, pwd, auto_disconnect = TRUE)

logout_openbis(token)

is_token_valid(token)
}
\arguments{
\item{user, pwd}{Login credentials for an openBis instance.}

\item{auto_disconnect}{Logical switch for automatically closing the
connection upon garbage collection of the token.}

\item{token}{Login token as created by \code{login_openbis()}.}
}
\description{
Create a login token for openBis API calls. If the \code{auto_disconnect} option
is chosen, the user is logged out automatically upon garbage collection of
the token. The validity of a token can be checked using \code{is_token_valid()}
and a login token can be destroyed using \code{logout_openbis()}.
}
\examples{
\donttest{
  tok <- login_openbis("rdgr2014", "IXPubReview", auto_disconnect = FALSE)
  is_token_valid(tok)
  logout_openbis(tok)
  is_token_valid(tok)
}

}
