% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_openbis}
\alias{login_openbis}
\alias{logout_openbis}
\alias{is_token_valid}
\title{Create and destroy a login token}
\usage{
login_openbis(user, pwd, auto_disconnect = TRUE, ...)

logout_openbis(token, ...)

is_token_valid(token, ...)
}
\arguments{
\item{user, pwd}{Login credentials for an openBis instance.}

\item{auto_disconnect}{Logical switch for automatically closing the
connection upon garbage collection of the token.}

\item{...}{Further arguments are forwarded to \code{\link[=make_request]{make_request()}}.}

\item{token}{Login token as created by \code{login_openbis()}.}
}
\description{
Login tokens for openBis API calls can be created using \code{login_openbis()}.
If the \code{auto_disconnect} option is enabled, the user is automatically
logged out using \code{logout_openbis()} upon garbage collection of the token.
The validity of a token can be checked using \code{is_token_valid()} and a login
token can be manually destroyed by calling \code{logout_openbis()} on the token.
}
\section{openBIS}{

\itemize{
\item \Sexpr{infx::docs_link("gis", "tryToAuthenticateForAllServices")}
}


\itemize{
\item \Sexpr{infx::docs_link("gis", "logout")}
}


\itemize{
\item \Sexpr{infx::docs_link("gis", "isSessionActive")}
}
}

\examples{
\donttest{
  # create a login token
  tok <- login_openbis("rdgr2014", "IXPubReview")
  # check token
  is_token_valid(tok)

  # destroy token
  logout_openbis(tok)
  # token is no longer valid
  is_token_valid(tok)
}

}
