% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{api_url}
\alias{api_url}
\alias{docs_link}
\title{OpenBIS urls}
\usage{
api_url(api_endpoint = c("gis", "gics", "qas", "wis", "dsrg", "sas", "dsrs"),
  host_url = "https://infectx.biozentrum.unibas.ch", full_url = NULL, ...)

docs_link(api_endpoint = c("gis", "gics", "qas", "wis", "dsrg", "sas",
  "dsrs"), method_name = NULL, version = "13.04.0")
}
\arguments{
\item{api_endpoint}{Abbreviated name of the API section (e.g. \code{gis} for
IGeneralInformationService).}

\item{host_url}{Host url.}

\item{full_url}{Instead of constructing the API endpoint url, a string can
be passed which will be returned again.}

\item{...}{Further arguments are ignored.}

\item{method_name}{Name of the method for which the link is created.}

\item{version}{OpenBIS version number.}
}
\description{
The helper function \code{api_url()} is used to create urls to InfectX
openBIS API endpoints and \code{docs_link()} generates latex links to
documentation pages. Both functions support the following endpoints which
are abbreviated as
\itemize{
\item \code{gis}: \Sexpr{infx::docs_link("gis")}
\item \code{gics}: \Sexpr{infx::docs_link("gics")}
\item \code{qas}: \Sexpr{infx::docs_link("qas")}
\item \code{wis}: \Sexpr{infx::docs_link("wis")}
\item \code{dsrg}: \Sexpr{infx::docs_link("dsrg")}
\item \code{sas}: \Sexpr{infx::docs_link("sas")}
\item \code{dsrs}: \Sexpr{infx::docs_link("dsrs")}
}
}
\details{
If for some reason an url is desired from \code{api_url()} that cannot be
constructed by pasting \code{host_url} and one of the hard-coded API endpoints
together, this can be passed as \code{full_url}, which will simply be returned.
}
\examples{
# default endpoint is the GeneralInformationService interface
api_url()
# base url can be customized
api_url(host_url = "https://foobar.xyz")
# ScreeningApiServer interface endpoint
api_url("sas")
# manual url
api_url(full_url = "https://foobar.xyz/openbis/new-api-section-v1.json")

# link to GeneralInformationService interface docs
docs_link()
# add a method name (only to the link text)
docs_link(method_name = "foo_bar")
# link to ScreeningApiServer interface docs
docs_link("sas")
# link to most recent version of docs
docs_link("sas", version = "16.05.6")

}
