% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{api_url}
\alias{api_url}
\alias{docs_link}
\title{OpenBIS urls}
\usage{
api_url(api = c("gis", "gics", "qas", "wis", "dsrg", "sas", "dsrs"),
  host = "https://infectx.biozentrum.unibas.ch")

docs_link(api = c("gis", "gics", "qas", "wis", "dsrg", "sas", "dsrs"),
  method_name = NULL, version = "13.04.0")
}
\arguments{
\item{api}{Abreviated name of the API section (e.g. \code{gis} for
IGeneralInformationService).}

\item{host}{Host url.}

\item{method_name}{Name of the method for which the link is created.}

\item{version}{OpenBIS version number.}
}
\description{
The helper function \code{api_url()} is used to create urls to InfectX
openBIS API endpoints and \code{docs_link()} generates latex links to
documentation pages. Both functions support the following endpoints which
are abbreviated as
\itemize{
\item \code{gis}: \Sexpr{infx::docs_link("gis")}
\item \code{gics}: \Sexpr{infx::docs_link("gics")}
\item \code{qas}: \Sexpr{infx::docs_link("qas")}
\item \code{wis}: \Sexpr{infx::docs_link("wis")}
\item \code{dsrg}: \Sexpr{infx::docs_link("dsrg")}
\item \code{sas}: \Sexpr{infx::docs_link("sas")}
\item \code{dsrs}: \Sexpr{infx::docs_link("dsrs")}
}
}
