% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mat_files}
\alias{read_mat_files}
\title{Read single cell data}
\usage{
read_mat_files(data)
}
\arguments{
\item{data}{The data to be read.}
}
\description{
Calling \code{\link[R.matlab:readMat]{R.matlab::readMat()}} single cell feature files in Matlab format, as
produced by CellProfiler are read and checked for the correct structure.
The Matlab file is expected to contain a nested list where each node
corresponds to an image and contains a list which is either holding a
single value or a vector of values.
}
