% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pub_meta}
\alias{read_pub_meta}
\title{Read public meta data}
\usage{
read_pub_meta(dat, ...)
}
\arguments{
\item{dat}{The data (raw vector) to be read.}

\item{...}{All further arguments are passed to [readr::read_delim].}
}
\value{
A tibble holding all read meta data or a subset thereof, depending
on the column specification.
}
\description{
Read public meta data downloaded from openBis using
[readr::read_delim].
}
