% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{make_requests}
\alias{make_requests}
\alias{make_request}
\alias{do_request_serial}
\alias{request_openbis}
\alias{api_url}
\alias{docs_link}
\alias{resolve_references}
\title{Make a JSON-RPC request}
\usage{
make_requests(urls, methods, params, names = NULL, ids = NULL,
  version = "2.0")

make_request(url, method, params, ...)

do_request_serial(urls, bodies, n_try = 2L, done = identity)

request_openbis(methods, params, api = c("IGeneralInformationService",
  "IGeneralInformationChangingService", "IQueryApiServer",
  "IWebInformationService", "IDssServiceRpcGeneric", "IScreeningApiServer",
  "IDssServiceRpcScreening"), host = "https://infectx.biozentrum.unibas.ch")

api_url(api = c("gis", "gics", "qas", "wis", "dsrg", "sas", "dsrs"),
  host = "https://infectx.biozentrum.unibas.ch")

docs_link(api = c("gis", "gics", "qas", "wis", "dsrg", "sas", "dsrs"),
  method_name = NULL, version = "13.04.0")

resolve_references(x)
}
\arguments{
\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The \code{@type} entries will be
generated from \code{json_class} attributes.}

\item{names}{Optional character vector of names to map list items to
api calls.}

\item{ids}{Identifier(s) for the JSON-RPC request (defaults to a random
string of length 7). Can be usually be ignored, as only single JSON-RPC
requests are issued per HTTP request.}

\item{version}{JSON-RPC protocol version to be used (defaults to \code{"2.0"}.}

\item{url, urls, }{Destination url(s), the request is sent to.}

\item{method, methods}{The API method name(s).}

\item{...}{Further arguments to \code{make_request} are passed to
\code{make_requests}.}

\item{bodies}{Request bodies: a list where each entry is a list with slots
\code{id}, \code{jsonrpc}, \code{method} and \code{params}.}

\item{n_try}{Number of tries each request is performed in case of failed
requests.}

\item{done}{A function that is applied to the result of a successful
request.}

\item{api, host}{Strings used to construct the destination url.}

\item{method_name}{Name of the method for which the link is created.}

\item{x}{A (possibly nested) list structure for which all \code{@id} fields are
recursively removed.}
}
\description{
Issues a POST request to a JSON-RPC server. All \code{@type} fields are
converted to/from \code{json_class} attributes, using \code{\link[=rm_json_class]{rm_json_class()}} and
\code{\link[=as_json_class]{as_json_class()}}. The helper function \code{request_openbis()} wraps
\code{make_requests()} and constructs the url the request is sent to, based on a
root url and an API section name (for the API section mapping, see
\href{https://wiki-bsse.ethz.ch/display/openBISDoc1304/openBIS+JSON+API}{docs}).
As part of the JSON-RPC specification, all objects returned form the API
will have \code{@id} fields, which may be referenced if an objects is used
multiple times. The helper function \code{resolve_references()} recursively
resolves all references such that each object is self-contained.
}
\examples{
\donttest{
  tok <- login_openbis("rdgr2014", "IXPubReview")
  projects <- request_openbis("listProjects", tok)
  print(projects[[1]])
}

}
