% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{make_request}
\alias{make_request}
\alias{request_openbis}
\alias{remove_id}
\title{Make a JSON-RPC request}
\usage{
make_request(url, method, params, version = "2.0", id = "1")

request_openbis(method, params, api = c("IGeneralInformationService",
  "IGeneralInformationChangingService", "IQueryApiServer",
  "IWebInformationService", "IDssServiceRpcGeneric", "IScreeningApiServer",
  "IDssServiceRpcScreening"), host = "https://infectx.biozentrum.unibas.ch")

remove_id(x)
}
\arguments{
\item{url}{Destination url, the request is sent to.}

\item{method}{The API method name.}

\item{params}{A list structure holding the arguments which, converted to
JSON, will be used to call the supplied method. The \code{@type} entries will be
generated from \code{json_class} attributes.}

\item{version}{JSON-RPC protocol version to be used (defaults to \code{"2.0"}.}

\item{id}{An identifier for the JSON-RPC request (defaults to \code{"1"}). Can be
ignored, as only single JSON-RPC requests are issued per HTTP request.}

\item{api, host}{Strings used to construct the destination url.}

\item{x}{A (possibly nested) list structure for which all \code{@id} fields are
recursively removed.}
}
\value{
A (nested) list holding the response from the JSON-RPC server
(\code{@type} entries are converted to \code{json_class} attributes).
}
\description{
Issues a POST request to a JSON-RPC server. All \code{@type} fields are
converted to/from \code{json_class} attributes, using \code{\link[=rm_json_class]{rm_json_class()}} and
\code{\link[=as_json_class]{as_json_class()}}. The helper function \code{request_openbis()} wraps
\code{make_request()} and constructs the url the request is sent to, based on a
root url and an API section name (for the API section mapping, see
\href{https://wiki-bsse.ethz.ch/display/openBISDoc1304/openBIS+JSON+API}{docs}).
As part of the JSON-RPC specification, all objects returned form the API
will have \code{@id} fields. The helper function \code{remove_id()} recursively
removes all fields with name \code{@id} from a list.
}
\examples{
\donttest{
  tok <- login_openbis("rdgr2014", "IXPubReview")
  projects <- request_openbis("listProjects", tok)
  print(projects[[1]])
}

}
